refactor: extract EmailImportService from TransactionService (740 -> 466 lines)
